Update getDayHourMinSecsForMillis to use long
This change does not affect existing implementations as
Java will natively promote int to long. Using long makes
more sense though, as time is usually expressed and stored
as a long in Java.

+++ src/main/java/com/hms_networks/americas/sc/extensions/system/time/SCTimeUtils.java
@@ -220,7 +220,7 @@ public class SCTimeUtils {
    *     [days text], [hours] [hours text], [minutes] [minutes text], [seconds] [seconds text]'
    */
   public static String getDayHourMinSecsForMillis(
-      int milliseconds, String daysText, String hoursText, String minutesText, String secondsText) {
+      long milliseconds, String daysText, String hoursText, String minutesText, String secondsText) {
     final long days = SCTimeUnit.MILLISECONDS.toDays(milliseconds);
     final long hours =
         SCTimeUnit.MILLISECONDS.toHours(milliseconds)
